Split acceptance tests up
